move ConfirmationPrompt to helpers; add a prompt when no assigned SEC
